Split BSON functionality into its own library.
BSON turns out to be a convenient mutable representation of state trees.
It can be used outside the context of MongoDB.
