+++ bosk-bson/README.md
@@ -1,0 +1,7 @@
+## bosk-bson
+
+This is the subproject for the published `bosk-bson` library,
+containing facilities for manipulating MongoDB's binary JSON data structures.
+Outside of MongoDB itself, we also use BSON as a handy utility
+for divide up and re-combine large JSON documents,
+which is useful even outside of MongoDB.
