+++ bosk-bson/build.gradle
@@ -1,0 +1,47 @@
+
+plugins {
+    id 'bosk.development'
+    id 'bosk.maven-publish'
+    id 'info.solidsoft.pitest' version '1.15.0'
+    id 'com.github.spotbugs' version '5.1.5'
+}
+
+base {
+    archivesName = 'bosk-bson'
+}
+
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(jdkVersion)
+    }
+}
+
+compileJava {
+    options.release = prodJavaVersion
+}
+
+compileTestJava {
+    options.release = null
+}
+
+dependencies {
+    api project(":bosk-core")
+    api 'org.mongodb:bson:5.1.2'
+    implementation 'com.github.spotbugs:spotbugs-annotations:4.8.6' // To stop warnings about When from MongoDB driver
+
+    testImplementation project(":bosk-logback")
+    testImplementation project(":bosk-testing")
+    testImplementation project(":lib-testing")
+}
+
+pitest {
+    pitestVersion = '1.15.0'
+    junit5PluginVersion = '1.2.0'
+    jvmArgs = ['-ea'] // Our unit tests check for assert statements
+    targetClasses = ['works.bosk.drivers.mongo.BsonSurgeon']
+    targetTests   = ['works.bosk.drivers.mongo.BsonSurgeonTest']
+    threads = 4
+    outputFormats = ['XML', 'HTML']
+    timestampedReports = false
+    //verbose = true
+}
