+++ bosk-mongo/src/main/java/works/bosk/drivers/mongo/Formatter.java
@@ -3,49 +3,32 @@ package works.bosk.drivers.mongo;
 import com.mongodb.client.model.changestream.ChangeStreamDocument;
 import com.mongodb.client.model.changestream.UpdateDescription;
 import java.lang.reflect.Type;
-import java.net.URLDecoder;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.NoSuchElementException;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.function.UnaryOperator;
-import java.util.regex.Pattern;
 import lombok.NonNull;
 import org.bson.BsonBinaryWriter;
 import org.bson.BsonDocument;
 import org.bson.BsonDocumentReader;
-import org.bson.BsonDocumentWriter;
 import org.bson.BsonInt32;
 import org.bson.BsonInt64;
-import org.bson.BsonReader;
 import org.bson.BsonString;
 import org.bson.BsonValue;
 import org.bson.codecs.BsonValueCodec;
 import org.bson.codecs.Codec;
 import org.bson.codecs.DecoderContext;
-import org.bson.codecs.DocumentCodecProvider;
 import org.bson.codecs.EncoderContext;
-import org.bson.codecs.ValueCodecProvider;
-import org.bson.codecs.configuration.CodecRegistries;
-import org.bson.codecs.configuration.CodecRegistry;
 import org.bson.io.BasicOutputBuffer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import works.bosk.BoskInfo;
-import works.bosk.Listing;
 import works.bosk.MapValue;
 import works.bosk.Reference;
 import works.bosk.SerializationPlugin;
-import works.bosk.SideTable;
-import works.bosk.exceptions.InvalidTypeException;
+import works.bosk.bson.BsonPlugin;
+import works.bosk.bson.BsonFormatter;
 
-import static java.lang.String.format;
 import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
 import static works.bosk.ReferenceUtils.rawClass;
@@ -55,10 +38,7 @@ import static works.bosk.ReferenceUtils.rawClass;
  *
  * @author pdoyle
  */
-final class Formatter {
-	private final CodecRegistry simpleCodecs;
-	private final Function<Type, Codec<?>> preferredBoskCodecs;
-	private final Function<Reference<?>, SerializationPlugin.DeserializationScope> deserializationScopeFunction;
+public final class Formatter extends BsonFormatter {
 
 	/**
 	 * If the diagnostic attributes are identical from one update to the next,
@@ -67,13 +47,8 @@ final class Formatter {
 	 */
 	private volatile MapValue<String> lastEventDiagnosticAttributes = MapValue.empty();
 
-	Formatter(BoskInfo<?> boskInfo, BsonPlugin bsonPlugin) {
-		this.simpleCodecs = CodecRegistries.fromProviders(
-			bsonPlugin.codecProviderFor(boskInfo),
-			new ValueCodecProvider(),
-			new DocumentCodecProvider());
-		this.preferredBoskCodecs = type -> bsonPlugin.getCodec(type, rawClass(type), simpleCodecs, boskInfo);
-		this.deserializationScopeFunction = bsonPlugin::newDeserializationScope;
+	public Formatter(BoskInfo<?> boskInfo, BsonPlugin bsonPlugin) {
+		super(boskInfo, bsonPlugin);
 	}
 
 	/**
@@ -99,44 +74,18 @@ final class Formatter {
 	 */
 	static final BsonInt64 REVISION_ONE = new BsonInt64(1);
 
-	/**
-	 * The fields of the main MongoDB document.  Case-sensitive.
-	 *
-	 * <p>
-	 * No field name should be a prefix of any other.
-	 */
-	enum DocumentFields {
-		path,
-		state,
-		revision,
-		diagnostics,
-	}
-
 	private final BsonInt32 SUPPORTED_MANIFEST_VERSION = new BsonInt32(1);
 
 	//
 	// Helpers to translate Bosk <-> MongoDB
 	//
 
-	Codec<?> codecFor(Type type) {
-		// BsonPlugin gives better codecs than CodecRegistry, because BsonPlugin is aware of generics,
-		// so we always try that first. The CodecSupplier protocol uses "null" to indicate that another
-		// CodecSupplier should be used, so we follow that protocol and fall back on the CodecRegistry.
-		// TODO: Should this logic be in BsonPlugin? It has nothing to do with MongoDriver really.
-		Codec<?> result = preferredBoskCodecs.apply(type);
-		if (result == null) {
-			return simpleCodecs.get(rawClass(type));
-		} else {
-			return result;
-		}
-	}
-
 	@SuppressWarnings("unchecked")
 	<T> T document2object(BsonDocument doc, Reference<T> target) {
 		Type type = target.targetType();
 		Class<T> objectClass = (Class<T>) rawClass(type);
 		Codec<T> objectCodec = (Codec<T>) codecFor(type);
-		try (@SuppressWarnings("unused") BsonPlugin.DeserializationScope scope = deserializationScopeFunction.apply(target)) {
+		try (@SuppressWarnings("unused") SerializationPlugin.DeserializationScope scope = deserializationScopeFunction.apply(target)) {
 			return objectClass.cast(objectCodec.decode(doc.asBsonReader(), DecoderContext.builder().build()));
 		}
 	}
@@ -200,42 +149,6 @@ final class Formatter {
 			result = result.with(name, value);
 		}
 		return result;
-	}
-
-	/**
-	 * @see #bsonValue2object(BsonValue, Reference)
-	 */
-	@SuppressWarnings("unchecked")
-	<T> BsonValue object2bsonValue(T object, Type type) {
-		rawClass(type).cast(object);
-		Codec<T> objectCodec = (Codec<T>) codecFor(type);
-		BsonDocument document = new BsonDocument();
-		try (BsonDocumentWriter writer = new BsonDocumentWriter(document)) {
-			// To support arbitrary values, not just whole documents, we put the result INSIDE a document.
-			writer.writeStartDocument();
-			writer.writeName("value");
-			objectCodec.encode(writer, object, EncoderContext.builder().build());
-			writer.writeEndDocument();
-		}
-		return document.get("value");
-	}
-
-	/**
-	 * @see #object2bsonValue(Object, Type)
-	 */
-	@SuppressWarnings("unchecked")
-	<T> T bsonValue2object(BsonValue bson, Reference<T> target) {
-		Codec<T> objectCodec = (Codec<T>) codecFor(target.targetType());
-		BsonDocument document = new BsonDocument();
-		document.append("value", bson);
-		try (
-			@SuppressWarnings("unused") BsonPlugin.DeserializationScope scope = deserializationScopeFunction.apply(target);
-			BsonReader reader = document.asBsonReader()
-		) {
-			reader.readStartDocument();
-			reader.readName("value");
-			return objectCodec.decode(reader, DecoderContext.builder().build());
-		}
 	}
 
 	@SuppressWarnings("unchecked")
@@ -309,162 +222,6 @@ final class Formatter {
 		} else {
 			lastEventDiagnosticAttributes = fromEvent;
 			return fromEvent;
-		}
-	}
-
-	/**
-	 * @return MongoDB field name corresponding to the given Reference
-	 * @see #referenceTo(String, Reference)
-	 */
-	static <T> String dottedFieldNameOf(Reference<T> ref, Reference<?> startingRef) {
-		return dottedFieldNameOf(ref, ref.path().length(), startingRef);
-	}
-
-	/**
-	 * @param refLength behave as though <code>ref</code> were truncated to this length, without actually having to do it
-	 * @return MongoDB field name corresponding to the given Reference
-	 * @see #referenceTo(String, Reference)
-	 */
-	static <T> String dottedFieldNameOf(Reference<T> ref, int refLength, Reference<?> startingRef) {
-		ArrayList<String> segments = dottedFieldNameSegments(ref, refLength, startingRef);
-		return String.join(".", segments.toArray(new String[0]));
-	}
-
-	static <T> ArrayList<String> dottedFieldNameSegments(Reference<T> ref, int refLength, Reference<?> startingRef) {
-		assert startingRef.path().matchesPrefixOf(ref.path()): "'" + ref + "' must be under '" + startingRef + "'";
-		ArrayList<String> segments = new ArrayList<>();
-		segments.add(DocumentFields.state.name());
-		buildDottedFieldNameOf(ref, startingRef.path().length(), refLength, segments);
-		return segments;
-	}
-
-	/**
-	 * @param elementRefLength behave as though <code>elementRef</code> were truncated to this length, without actually having to do it
-	 * @return MongoDB field name corresponding to the container (Catalog or SideTable) that contains the given element
-	 * @see #referenceTo(String, Reference)
-	 */
-	static <T> List<String> containerSegments(Reference<T> elementRef, int elementRefLength, Reference<?> startingRef) {
-		List<String> elementSegments = dottedFieldNameSegments(elementRef, elementRefLength, startingRef);
-		return elementSegments.subList(0, elementSegments.size()-1); // Trim off the element itself
-	}
-
-	/**
-	 * @param refLength behave as though <code>ref</code> were truncated to this length without actually having to do it
-	 */
-	private static <T> void buildDottedFieldNameOf(Reference<T> ref, int startLength, int refLength, ArrayList<String> segments) {
-		if (ref.path().length() > startLength) {
-			Reference<?> enclosingReference = ref.enclosingReference(Object.class);
-			buildDottedFieldNameOf(enclosingReference, startLength, refLength, segments);
-			if (ref.path().length() <= refLength) {
-				if (Listing.class.isAssignableFrom(enclosingReference.targetClass())) {
-					segments.add("ids");
-				} else if (SideTable.class.isAssignableFrom(enclosingReference.targetClass())) {
-					segments.add("valuesById");
-				}
-				segments.add(dottedFieldNameSegment(ref.path().lastSegment()));
-			}
-		}
-	}
-
-	static String dottedFieldNameSegment(String segment) {
-		return ENCODER.apply(segment);
-	}
-
-	static String undottedFieldNameSegment(String dottedSegment) {
-		return DECODER.apply(dottedSegment);
-	}
-
-	/**
-	 * @return Reference corresponding to the given field name
-	 * @see #dottedFieldNameOf
-	 */
-	@SuppressWarnings("unchecked")
-	static <T> Reference<T> referenceTo(String dottedName, Reference<?> startingReference) throws InvalidTypeException {
-		Reference<?> ref = startingReference;
-		Iterator<String> iter = Arrays.asList(dottedName.split(Pattern.quote("."))).iterator();
-		skipField(ref, iter, DocumentFields.state.name()); // The entire Bosk state is in this field
-		while (iter.hasNext()) {
-			if (Listing.class.isAssignableFrom(ref.targetClass())) {
-				skipField(ref, iter, "ids");
-			} else if (SideTable.class.isAssignableFrom(ref.targetClass())) {
-				skipField(ref, iter, "valuesById");
-			}
-			if (iter.hasNext()) {
-				String segment = undottedFieldNameSegment(iter.next());
-				ref = ref.then(Object.class, segment);
-			}
-		}
-		return (Reference<T>) ref;
-	}
-
-	private static void skipField(Reference<?> ref, Iterator<String> iter, String expectedName) {
-		String actualName;
-		try {
-			actualName = iter.next();
-		} catch (NoSuchElementException e) {
-			throw new IllegalStateException("Expected '" + expectedName + "' for " + ref.targetClass().getSimpleName() + "; encountered end of dotted field name");
-		}
-		if (!expectedName.equals(actualName)) {
-			throw new IllegalStateException("Expected '" + expectedName + "' for " + ref.targetClass().getSimpleName() + "; was: " + actualName);
-		}
-	}
-
-	private static final UnaryOperator<String> DECODER;
-	private static final UnaryOperator<String> ENCODER;
-
-	static {
-		DECODER = s->{
-			return URLDecoder.decode(s, StandardCharsets.UTF_8);
-		};
-
-		ENCODER = s->{
-			// Selective percent-encoding of characters MongoDB doesn't like.
-			// Standard percent-encoding doesn't handle the period character, which
-			// we want, so if we're already diverging from the standard, we might
-			// as well do something that suits our needs.
-			// Good to stay compatible with standard percent-DEcoding, though.
-			StringBuilder sb = new StringBuilder();
-			for (int i = 0; i < s.length(); ) {
-				int cp = s.codePointAt(i);
-				switch (cp) {
-					case '%': // For percent-encoding
-					case '+': case ' ': // These two are affected by URLDecoder
-					case '$': // MongoDB treats these specially
-					case '.': // MongoDB separator for dotted field names
-					case 0:   // Can MongoDB handle nulls? Probably. Do we want to find out? Not really.
-					case '|': // (These are reserved for internal use)
-					case '!':
-					case '~':
-					case '[':
-					case ']':
-						appendPercentEncoded(sb, cp);
-						break;
-					default:
-						sb.appendCodePoint(cp);
-						break;
-				}
-				i += Character.charCount(cp);
-			}
-			return sb.toString();
-		};
-	}
-
-	private static void appendPercentEncoded(StringBuilder sb, int cp) {
-		assert 0 <= cp && cp <= 255;
-		sb
-			.append('%')
-			.append(hexCharForDigit(cp / 16))
-			.append(hexCharForDigit(cp % 16));
-	}
-
-	/**
-	 * An uppercase version of {@link Character#forDigit} with a radix of 16.
-	 */
-	private static char hexCharForDigit(int value) {
-		if (value < 10) {
-			return (char)('0' + value);
-		} else {
-			return (char)('A' + value - 10);
 		}
 	}
