+++ bosk-mongo/src/main/java/works/bosk/drivers/mongo/MainDriver.java
@@ -29,7 +29,8 @@ import works.bosk.BoskInfo;
 import works.bosk.Identifier;
 import works.bosk.Reference;
 import works.bosk.StateTreeNode;
-import works.bosk.drivers.mongo.Formatter.DocumentFields;
+import works.bosk.bson.BsonPlugin;
+import works.bosk.bson.BsonFormatter.DocumentFields;
 import works.bosk.drivers.mongo.MappedDiagnosticContext.MDCScope;
 import works.bosk.drivers.mongo.MongoDriverSettings.DatabaseFormat;
 import works.bosk.drivers.mongo.MongoDriverSettings.InitialDatabaseUnavailableMode;
