+++ bosk-mongo/src/main/java/works/bosk/drivers/mongo/SequoiaFormatDriver.java
@@ -28,7 +28,8 @@ import works.bosk.Identifier;
 import works.bosk.MapValue;
 import works.bosk.Reference;
 import works.bosk.StateTreeNode;
-import works.bosk.drivers.mongo.Formatter.DocumentFields;
+import works.bosk.bson.BsonPlugin;
+import works.bosk.bson.BsonFormatter.DocumentFields;
 import works.bosk.exceptions.FlushFailureException;
 import works.bosk.exceptions.InvalidTypeException;
 
@@ -42,7 +43,7 @@ import static java.util.Objects.requireNonNull;
 import static org.bson.BsonBoolean.FALSE;
 import static works.bosk.drivers.mongo.Formatter.REVISION_ZERO;
 import static works.bosk.drivers.mongo.Formatter.dottedFieldNameOf;
-import static works.bosk.drivers.mongo.Formatter.referenceTo;
+import static works.bosk.bson.BsonFormatter.referenceTo;
 import static works.bosk.drivers.mongo.MainDriver.MANIFEST_ID;
 import static works.bosk.drivers.mongo.MongoDriverSettings.ManifestMode.CREATE_IF_ABSENT;
