+++ bosk-mongo/src/test/java/works/bosk/drivers/mongo/MiscMongoDriverTest.java
@@ -1,0 +1,14 @@
+package works.bosk.drivers.mongo;
+
+import org.junit.jupiter.api.Test;
+import works.bosk.exceptions.InvalidTypeException;
+
+import static works.bosk.TypeValidation.validateType;
+
+public class MiscMongoDriverTest {
+	@Test
+	void manifest_passesTypeValidation() throws InvalidTypeException {
+		validateType(Manifest.class);
+	}
+
+}
