+++ bosk-mongo/src/test/java/works/bosk/drivers/mongo/MongoDriverConformanceTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import works.bosk.DriverFactory;
 import works.bosk.StateTreeNode;
+import works.bosk.bson.BsonPlugin;
 import works.bosk.drivers.DriverConformanceTest;
 import works.bosk.drivers.mongo.TestParameters.EventTiming;
 import works.bosk.drivers.mongo.TestParameters.ParameterSet;
