+++ bosk-mongo/src/test/java/works/bosk/drivers/mongo/MongoDriverHanoiTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.TestInfo;
 import works.bosk.DriverStack;
+import works.bosk.bson.BsonPlugin;
 import works.bosk.drivers.HanoiTest;
 import works.bosk.junit.ParametersByName;
 import works.bosk.junit.Slow;
