+++ bosk-mongo/src/test/java/works/bosk/drivers/mongo/SchemaEvolutionTest.java
@@ -7,6 +7,7 @@ import java.util.stream.Stream;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,10 +15,12 @@ import works.bosk.Bosk;
 import works.bosk.Reference;
 import works.bosk.annotations.ReferencePath;
 import works.bosk.drivers.state.TestEntity;
+import works.bosk.exceptions.InvalidTypeException;
 import works.bosk.junit.ParametersByName;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static works.bosk.BoskTestUtils.boskName;
+import static works.bosk.TypeValidation.validateType;
 
 @UsesMongoService
 public class SchemaEvolutionTest {
