+++ lib-testing/build.gradle
@@ -18,7 +18,7 @@ dependencies {
 	implementation project(":bosk-testing")
 	// These are for AbstractRoundTripTest. That logic ought to be moved to their respective sub-projects
 	implementation project(":bosk-jackson")
-	implementation project(":bosk-mongo")
+	implementation project(":bosk-bson")
 
 	// The bosk.development plugin brings these in as test dependencies,
 	// but we need them as main dependencies.
