+++ settings.gradle
@@ -1,6 +1,7 @@
 
 rootProject.name = 'bosk'
 include 'bosk-annotations'
+include 'bosk-bson'
 include 'bosk-core'
 include 'bosk-jackson'
 include 'bosk-logback'
